+++ hs-ucan/test/Test/Web/UCAN/Prelude.hs
@@ -1,4 +1,4 @@
-module Test.Prelude
+module Test.Web.UCAN.Prelude
   ( module RIO
   , module Data.Aeson
